+++ changelog.md
@@ -5,6 +5,9 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add `click:source` directive as an alias to `click:example` directive in Sphinx extension.
+- Flag `click:example` directive as deprecated in favor of `click:source`.
+
 ## [7.2.0 (2025-11-26)](https://github.com/kdeldycke/click-extra/compare/v7.1.0...v7.2.0)
 
 - Add support for auto-conversion of GitHub alerts into MyST admonitions in Sphinx extension.